+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/ContributesSubcomponentHandlerGenerator.kt
@@ -294,7 +294,7 @@ internal class ContributesSubcomponentHandlerGenerator(
       )
     }
 
-    val functions = componentInterface.functions
+    val functions = componentInterface.memberFunctions
       .filter { it.isAbstract() && it.visibility() == PUBLIC }
       .filter {
         val returnType = it.returnType().asClassReference()
@@ -333,9 +333,14 @@ internal class ContributesSubcomponentHandlerGenerator(
           )
         }
 
-        val createComponentFunctions = factory.functions
+        val createComponentFunctions = factory.memberFunctions
           .filter { it.isAbstract() }
-          .filter { it.returnType().asClassReference().fqName == contributionFqName }
+          .filter {
+            val returnType = it.returnType().asClassReference()
+            returnType.fqName == contributionFqName ||
+              contribution.clazz.directSuperTypeReferences()
+                .any { type -> type.asClassReference() == returnType }
+          }
 
         if (createComponentFunctions.size != 1) {
           throw AnvilCompilationExceptionClassReference(