+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/KspContributesSubcomponentHandlerSymbolProcessor.kt
@@ -405,9 +405,14 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
               } else {
                 function.asMemberOf(implementingType).returnTypeOrNull()
               }
-            returnTypeToCheck
-              ?.resolveKSClassDeclaration()
-              ?.toClassName() == contributionClassName
+
+            if (returnTypeToCheck != null) {
+              val returnTypeClassName = returnTypeToCheck.resolveKSClassDeclaration()?.toClassName()
+              returnTypeClassName == contributionClassName ||
+                returnTypeToCheck.isAssignableFrom(contribution.clazz.asType(emptyList()))
+            } else {
+              false
+            }
           }
           .toList()